fix: cbuffer layout mismatch - hlgMode must use min16float
The HLSL packing rules do not allow mixed-precision types to share the
same 16-byte constant buffer register. Declaring hlgMode as 'float'
after 'min16float3 offsets' caused hlgMode to be placed at offset 64
(new register) instead of offset 60 (packing into the offsets register).

This shifted chromaOffset to 80 and chromaTexMax to 88, mismatching
the C++ CSC_CONST_BUF layout where they reside at offsets 64 and 72.
The resulting misaligned reads corrupted chroma sampling parameters,
causing a blue color tint across ALL rendering modes (SDR, PQ, HLG).

Fix: change hlgMode from 'float' to 'min16float' so it packs at
offset 60 alongside offsets, matching the C++ struct layout.

+++ app/shaders/d3d11_yuv420_pixel.hlsl
@@ -12,7 +12,7 @@ cbuffer CSC_CONST_BUF : register(b0)
 {
     min16float3x3 cscMatrix;
     min16float3 offsets;
-    float hlgMode;
+    min16float hlgMode;
     min16float2 chromaOffset;
     min16float2 chromaTexMax;
 };
